requirements.txt
torch==2.4.1
numpy>=1.24.0
opencv-python>=4.8.0
h5py>=3.0.0
wandb>=0.15.0
tqdm>=4.65.0
hydra-core>=1.3.0
pillow>=10.0.0
mlagents==1.1.0
mlagents-envs==1.1.0
gym>=0.26.0
pytest>=7.0.0
black>=23.0.0
isort>=5.12.0
